src/test/java/com/example/mylittleplayer/PlaylistTest.java
package com.example.mylittleplayer;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.util.ArrayList;
/* Copyright (C) 2022 <Название команды>
   This file is part of the MediaPlayer.
   (Yet Another Merge Sort Routines)

   MediaPlayer is free software: you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
    the Free Software Foundation, either version 3 of the License, or
    (at your option) any later version.

    MediaPlayer is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
    GNU General Public License for more details.

    You should have received a copy of the GNU General Public License
    along with MediaPlayer.  If not, see <http://www.gnu.org/licenses/>.

  (Этот файл — часть MediaPlayer.

   MediaPlayer - свободная программа: вы можете перераспространять ее и/или
   изменять ее на условиях Стандартной общественной лицензии GNU в том виде,
   в каком она была опубликована Фондом свободного программного обеспечения;
   либо версии 3 лицензии, либо (по вашему выбору) любой более поздней
   версии.

   MediaPlayer распространяется в надежде, что она будет полезной,
   но БЕЗО ВСЯКИХ ГАРАНТИЙ; даже без неявной гарантии ТОВАРНОГО ВИДА
   или ПРИГОДНОСТИ ДЛЯ ОПРЕДЕЛЕННЫХ ЦЕЛЕЙ. Подробнее см. в Стандартной
   общественной лицензии GNU.

   Вы должны были получить копию Стандартной общественной лицензии GNU
   вместе с этой программой. Если это не так, см.
   <http://www.gnu.org/licenses/>.)

 */
class PlaylistTest {

    Playlist playlist = new Playlist();

    @BeforeEach
    void setUp() {
    }

    @AfterEach
    void tearDown() {
    }

    @Test
    void getName() {
        playlist.setName("Name");
        assert("Name".equals( playlist.getName()));
    }

    @Test
    void setName() {
        playlist.setName("Name");
        assert(!(playlist.getName().equals("")));
    }

    @Test
    void getSongs() {
        assert(playlist.getSongs().equals(new ArrayList<Song>()));
    }

    @Test
    void setSongs() {
    }
}